Prepare release 0.5.0
