+++ src/DataTypeFactory.php
@@ -6,6 +6,8 @@ use OutOfBoundsException;
 use InvalidArgumentException;
 
 /**
+ * @since 0.1
+ *
  * @licence GNU GPL v2+
  * @author Daniel Kinzler
  */
